added CommandOptionExtensions, little clean up, main entry point to return int
